+++ nextjs-metrics-viewer/app/dashboard/page.tsx
@@ -10,7 +10,7 @@ const DynamicMetricsSection = dynamic(
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center gap-8 bg-black py-32">
+    <main className="flex min-h-screen flex-col items-center gap-8 bg-black px-6 py-32">
       <Header />
       <h1 className="text-2xl text-white">Dashboard</h1>
